[FIX] Take into account reconciled invoice journal item - Issue12013
Issue
-
https://www.vauxoo.com/web#id=12013&model=helpdesk.ticket&view_type=form&menu_id=

Main
-

Take into account Invoices that have been fully paid though their
reconciliation chain is not fully reconciled.

Explanation
-

- Happy path: One invoice is paid by One paid no residual left.
- Happy path: Several Invoices are fully paid by one Single Payment and
there are no leftovers.

- This path: There are cases when several invoices are paid by several
payments/credit notes but the reconciliation in whole chain is not
performed in full. Thus this case is left with one invoice that was fully
paid but no full-reconciliation is performed and no commission computed.